Moved corkscrew backtrace to own file
